Create the tools section and refactor code
