+++ resources/js/console/Scene.tsx
@@ -19,6 +19,7 @@ import {hasTrialEndedAndNotSubscribed} from "./lib/blog-helpers";
 import BlogTrialEnded from "./Views/BlogTrialEnded";
 import Integrations from "./Integrations/Integrations";
 import { router } from 'kea-router'
+import Tools from './Tools/Tools'
 
 export const scenes = {
     error404: () => <div>404</div>,
@@ -26,7 +27,7 @@ export const scenes = {
     posts: ({ postId } : { postId?: number }) => <Posts postId={postId} />,
     pages: ({ postId } : { postId?: number }) => <Pages postId={postId} />,
     settings: ({type} : {type?: string}) => <Settings type={type} />,
-    tools: ({type} : {type?: string}) => <Settings type={type} />,
+    tools: ({type} : {type?: string}) => <Tools type={type} />,
     integrations: ({type} : {type?: string}) => <Integrations type={type} />,
     theme: ({type} : {type?: string }) => <Theme />,
     billing: () => <Billing />,
